split simulate linkedreads and variants for sanity reasons
